requirements.txt
Django>=3.2,<4
djangorestframework>=3.14.0,<4
requests>=2.28.1,<3
drf-yasg>=1.21.7,<2
celery[redis]>=5.2.7,<6
importlib-metadata==4.13.0
pytest>=8.0.0,<9
pytest-django>=4.8.0,<5